refactor(frontend): Reorganizing components folder
